キャッシュ保存フォルダが選択されていない場合に落ちる問題に対応

+++ NicoPlayerHohoema/Models/Video/DividedQualityNicoVideo.cs
@@ -302,6 +302,7 @@ namespace NicoPlayerHohoema.Models
 		protected Task DeleteDownloadProgress()
 		{
 			if (!IsAvailable) { return Task.CompletedTask; }
+			if (_DownloadProgressFileAccessor == null) { return Task.CompletedTask; }
 
 			return _DownloadProgressFileAccessor.Delete();
 		}
